Refactor interpreter
Extract the general purpose logic from the splice interpreter. The new interpreter
class will be the basis for the macro annotation interpreter. The splice interpreter
only keeps logic related with level -1 quote and type evaluation.

Part of https://github.com/dotty-staging/dotty/tree/design-macro-annotations
